Git Sparse Checkout Testing
I managed to maybe get sparse checkout to work but it needs a lot of testing. 

Also I found out that PATH can be ~/bin/win-sudo instead of /c/Users/win-sudo. The problem is that the script doesn't consider '/c' to be a valid directory so it needs a rework. 

The commented 'git clone' lines work fine with this script and I managed to do a successful install using my fork!